finish helper tests

+++ src/helpers/clearTmpFile.ts
@@ -1,5 +1,8 @@
 import fs from 'fs';
 
 export const clearTmpFile = (filePath: string): void => {
+  if (!fs.existsSync(filePath)) {
+    throw new Error('File does not exist');
+  }
   fs.unlinkSync(filePath);
 };
